Create Awareness Report

This function generates an awareness report in HTML format, and optionally opens it in the web browser.

Usage

create_awareness_report(
  DD.dict,
  DS.data,
  non.NA.missing.codes = NA,
  threshold = 95,
  output.path = tempdir(),
  open.html = TRUE,
  fn.stem = "AwarenessReport"
)

Arguments

DD.dict

Data dictionary.

DS.data

Data set.

non.NA.missing.codes

A user-defined vector of numerical missing value codes (e.g., -9999).

threshold

Threshold for missingness of concern (as a percent).

output.path

Path to the folder in which to create the HTML report document.

open.html

If TRUE, open the HTML report document in the web browser.

fn.stem

File name stem.

Value

Full path to the HTML report document.

See also

Examples

# \donttest{
data(ExampleB)
create_awareness_report(DD.dict.B, DS.data.B, non.NA.missing.codes=c(-9999),
   output.path= tempdir(), open.html = FALSE)
#> 
#> 
#> processing file: dbGaP_awareness_report.Rmd
#> 1/18           
#> 2/18 [su]      
#> 3/18           
#> 4/18 [param]   
#> 5/18           
#> 6/18 [misssum] 
#> 7/18           
#> 8/18 [vmt1a]   
#> 9/18           
#> 10/18 [vmt2b]   
#> 11/18           
#> 12/18 [vmt3c]   
#> 13/18           
#> 14/18 [vmt4d]   
#> 15/18           
#> 16/18 [vmt4e]   
#> 17/18           
#> 18/18 [sess_inf]
#> output file: dbGaP_awareness_report.knit.md
#> /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/pandoc +RTS -K512m -RTS dbGaP_awareness_report.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/AwarenessReport-20250501_100635.html --lua-filter /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --variable bs3=TRUE --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/rmarkdown-strd2c024467505.html 
#> 
#> Output created: /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/AwarenessReport-20250501_100635.html
#> [1] "/var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/AwarenessReport-20250501_100635.html"
# }