Create Report

This function calls eval_function to generate a textual and graphical report of the selected variables in HTML format, and optionally opens it in the web browser.

Usage

create_report(
  DD.dict,
  DS.data,
  sex.split = FALSE,
  sex.name = NULL,
  start = 1,
  end = 1,
  non.NA.missing.codes = NA,
  output.path = tempdir(),
  open.html = TRUE,
  fn.stem = "Report"
)

Arguments

DD.dict

Data dictionary.

DS.data

Data set.

sex.split

When TRUE, split reports by the field named as defined by the sex.name variable.

sex.name

Character string specifying the name of the sex field.

start

Staring index of the first select trait.

end

Ending index of the last selected trait.

non.NA.missing.codes

A user-defined vector of numerical missing value codes (e.g., -9999).

output.path

Path to the folder in which to create the HTML report document.

open.html

If TRUE, open the HTML report document in the web browser.

fn.stem

File name stem.

Value

Full path to the HTML report document.

Examples

# \donttest{
data(ExampleB)
create_report(DD.dict.B, DS.data.B, sex.split=TRUE, sex.name= "SEX",
   start = 3, end = 7, non.NA.missing.codes=c(-9999,-4444),
   output.path= tempdir(), open.html = FALSE)
#> 
#> 
#> processing file: dbGaP_check_report.Rmd
#> 1/11             
#> 2/11 [set]       
#> 3/11             
#> 4/11 [ll]        
#> 5/11             
#> 6/11 [parameters]
#> 7/11             
#> 8/11 [applyfun5] 
#> 9/11             
#> 10/11 [si]        
#> 11/11             
#> output file: dbGaP_check_report.knit.md
#> /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/pandoc +RTS -K512m -RTS dbGaP_check_report.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/Report-20250501_100635.html --lua-filter /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --variable bs3=TRUE --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/rmarkdown-strd2c049e69125.html 
#> 
#> Output created: /var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/Report-20250501_100635.html
#> [1] "/var/folders/y1/p2tdt78d16b191mrhv6hsm0c0000gq/T//Rtmpktojw4/Report-20250501_100635.html"
# }