Min Max Required Pre-checks

This function runs a workflow of the minimum number of checks required for a user to run minmax_check; the checks include pkg_field_check, dimension_check, and name_check.

Usage

mm_precheck(dict, data)

Arguments

dict

Data dictionary.

data

Data set.

Value

Tibble containing the following information for each check: (1) Time (time stamp); (2) Name (name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (More detailed information about the potential errors identified).

Examples

data(ExampleB)
mm_precheck(DD.dict.B, DS.data.B)
#> # A tibble: 3 × 5
#>   Time                Function        Status Message                 Information
#>   <dttm>              <chr>           <chr>  <chr>                   <named lis>
#> 1 2025-05-01 10:29:09 pkg_field_check Passed Passed: package-level … <lgl [3]>  
#> 2 2025-05-01 10:29:09 dimension_check Passed Passed: the variable c… <int [2]>  
#> 3 2025-05-01 10:29:09 name_check      Passed Passed: the variable n… <chr [1]>