Miscellaneous Format Check

This function checks miscellaneous dbGaP formatting requirements to ensure (1) no empty variable names; (2) no duplicate variable names; (3) variable names do not contain "dbgap"; (4) there are no duplicate column names in the dictionary; and (5) column names falling after VALUES column are unnamed.

Usage

misc_format_check(DD.dict, DS.data, verbose = TRUE)

Arguments

DD.dict

Data dictionary.

DS.data

Data set.

verbose

When TRUE, the function prints the Message out, as well as more detailed information about which formatting checks failed.

Value

Tibble, returned invisibly, containing: (1) Time (time stamp); (2) Name (name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (Names of variables that fail one of these checks).

Examples

# Example 1: Fail check 
data(ExampleJ)
misc_format_check(DD.dict.J, DS.data.J)
#> $Message
#> [1] "ERROR: at least one check failed."
#> 
#> $Information
#> # A tibble: 6 × 4
#>   check.name check.description                           check.status details   
#>   <chr>      <chr>                                       <chr>        <chr>     
#> 1 Check 1    Empty variable name check                   Passed       NA        
#> 2 Check 2    Duplicate variable name check               Passed       NA        
#> 3 Check 3    Check for use of `dbgap` in variable names  Failed       HTN_dbGaP 
#> 4 Check 3    Check for use of `dbgap` in variable names  Failed       PHYSICAL_…
#> 5 Check 4    Duplicate dictionary column name check      Passed       NA        
#> 6 Check 5    Column names after `VALUES` should be empty Passed       NA        
#> 
print(misc_format_check(DD.dict.J, DS.data.J, verbose=FALSE))
#> # A tibble: 1 × 5
#>   Time                Function          Status Message                  Inform…¹
#>   <dttm>              <chr>             <chr>  <chr>                    <named >
#> 1 2023-02-13 16:52:54 misc_format_check Failed ERROR: at least one che… <tibble>
#> # … with abbreviated variable name ¹​Information

# Example 2: Pass check
data(ExampleA)
misc_format_check(DD.dict.A, DS.data.A)
#> $Message
#> [1] "Passed: no check-specific formatting issues identified."
#> 
#> $Information
#> # A tibble: 5 × 4
#>   check.name check.description                           check.status details
#>   <chr>      <chr>                                       <chr>        <lgl>  
#> 1 Check 1    Empty variable name check                   Passed       NA     
#> 2 Check 2    Duplicate variable name check               Passed       NA     
#> 3 Check 3    Check for use of `dbgap` in variable names  Passed       NA     
#> 4 Check 4    Duplicate dictionary column name check      Passed       NA     
#> 5 Check 5    Column names after `VALUES` should be empty Passed       NA     
#> 
print(misc_format_check(DD.dict.A, DS.data.A, verbose=FALSE))
#> # A tibble: 1 × 5
#>   Time                Function          Status Message                  Inform…¹
#>   <dttm>              <chr>             <chr>  <chr>                    <named >
#> 1 2023-02-13 16:52:54 misc_format_check Passed Passed: no check-specif… <tibble>
#> # … with abbreviated variable name ¹​Information